+++ core/templates/dev/head/pages/exploration_editor/ExplorationEditor.ts
@@ -115,7 +115,7 @@ require('pages/exploration_editor/feedback_tab/ThreadTableDirective.ts');
 require('pages/exploration_editor/history_tab/HistoryTab.ts');
 require(
   'pages/exploration_editor/improvements_tab/ImprovementsTabDirective.ts');
-require('pages/exploration_editor/preview_tab/PreviewTab.ts');
+require('pages/exploration_editor/preview_tab/PreviewTabDirective.ts');
 require('pages/exploration_editor/settings_tab/SettingsTab.ts');
 require('pages/exploration_editor/statistics_tab/PieChartDirective.ts');
 require(
